Refactoring to move behaviour that validate NotFound agenda by id to VoteSession domain
