+++ src/test/java/com/cooperative/assembly/voting/meeting/agenda/MeetingAgendaServiceTest.java
@@ -10,6 +10,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import java.util.Optional;
+
+import static java.lang.Boolean.TRUE;
 import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
@@ -17,6 +20,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasProperty;
 import static org.hamcrest.Matchers.isEmptyString;
+import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.mockito.Mockito.*;
 import static java.util.Optional.of;
@@ -82,29 +86,10 @@ public class MeetingAgendaServiceTest {
         MeetingAgenda expectedAgenda = new MeetingAgenda(agendaId, "agenda-title-1");
         when(repository.findById(agendaId)).thenReturn(of(expectedAgenda));
 
-        MeetingAgenda agenda = service.loadAgenda(agendaId);
+        Optional<MeetingAgenda> agenda = service.loadAgenda(agendaId);
 
-        assertThat(agenda, hasProperty("id", equalTo(expectedAgenda.getId())));
-    }
-
-    @Test
-    public void shouldThrowNotFoundReferenceExceptionExceptionWhenCanNotFindAnyAgendaByGivenId() {
-        String agendaId = randomUUID().toString();
-        when(repository.findById(agendaId)).thenReturn(empty());
-
-        assertThatExceptionOfType(NotFoundReferenceException.class)
-                .isThrownBy(() -> service.loadAgenda(agendaId))
-                .withMessage("Reference not found");
-    }
-
-    @Test
-    public void shouldThrowNotFoundReferenceExceptionExceptionWhenCanFindAnAgendaByGivenId() {
-        String agendaId = randomUUID().toString();
-        MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
-        when(repository.findById(agendaId)).thenReturn(of(agenda));
-
-        assertThatCode(() -> service.loadAgenda(agendaId))
-                .doesNotThrowAnyException();
+        assertThat(agenda.isPresent(), is(TRUE));
+        assertThat(agenda.get(), hasProperty("id", equalTo(expectedAgenda.getId())));
     }
 
 }
