+++ src/test/java/com/cooperative/assembly/voting/vote/session/VoteSessionServiceTest.java
@@ -1,5 +1,6 @@
 package com.cooperative.assembly.voting.vote.session;
 
+import com.cooperative.assembly.voting.error.exception.NotFoundReferenceException;
 import com.cooperative.assembly.voting.error.exception.ValidationException;
 import com.cooperative.assembly.voting.meeting.agenda.MeetingAgenda;
 import com.cooperative.assembly.voting.meeting.agenda.MeetingAgendaService;
@@ -14,9 +15,11 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
+import static java.util.Optional.empty;
 import static java.util.UUID.randomUUID;
 import static java.time.LocalDateTime.now;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -24,6 +27,7 @@ import static org.hamcrest.Matchers.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
+import static java.util.Optional.of;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
@@ -60,6 +64,9 @@ public class VoteSessionServiceTest {
 
     @Test
     public void shouldSaveVoteSessionWhenOpeningVoteSessionForGivenAgendaId() {
+        MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(agenda));
+
         service.openFor(agendaId, deadlineMinutes);
 
         verify(repository, times(1)).save(any(VoteSession.class));
@@ -67,6 +74,9 @@ public class VoteSessionServiceTest {
 
     @Test
     public void shouldLoadAgendaByGivenAgendaIdWhenOpeningVoteSession() {
+        MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(agenda));
+
         service.openFor(agendaId, deadlineMinutes);
 
         verify(meetingAgendaService, times(1)).loadAgenda(agendaId);
@@ -75,7 +85,7 @@ public class VoteSessionServiceTest {
     @Test
     public void shouldSaveVoteSessionWithUUIDAndOpeningAndClosingTimeAndAgendaLoadedById() {
         MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
-        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(agenda);
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(agenda));
 
         service.openFor(agendaId, deadlineMinutes);
 
@@ -90,7 +100,7 @@ public class VoteSessionServiceTest {
     @Test
     public void shouldReturnSavedVoteSessionAsIs() {
         MeetingAgenda expectedAgenda = new MeetingAgenda(agendaId, "agenda-title-1");
-        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(expectedAgenda);
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(expectedAgenda));
 
         VoteSession expectedVoteSession = new VoteSession(sessionId, expectedAgenda, openingTime, closingTime);
         when(repository.save(any(VoteSession.class))).thenReturn(expectedVoteSession);
@@ -118,10 +128,30 @@ public class VoteSessionServiceTest {
     @Test
     public void shouldNeverThrowsAnyNotFoundReferenceExceptionExceptionWhenCanNotFindAnyAgendaByGivenId() {
         MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
-        VoteSession foundSession = new VoteSession(sessionId, agenda, openingTime, closingTime);
         when(repository.findByAgendaId(agendaId)).thenReturn(emptyList());
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(agenda));
 
         assertThatCode(() -> service.openFor(agendaId, deadlineMinutes))
+                .doesNotThrowAnyException();
+    }
+
+    @Test
+    public void shouldThrowNotFoundReferenceExceptionExceptionWhenCanNotFindAnyAgendaByGivenId() {
+        String agendaId = randomUUID().toString();
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(empty());
+
+        assertThatExceptionOfType(NotFoundReferenceException.class)
+                .isThrownBy(() -> service.loadSessionAgenda(agendaId))
+                .withMessage("Reference not found");
+    }
+
+    @Test
+    public void shouldThrowNotFoundReferenceExceptionExceptionWhenCanFindAnAgendaByGivenId() {
+        String agendaId = randomUUID().toString();
+        MeetingAgenda agenda = new MeetingAgenda(agendaId, "agenda-title-1");
+        when(meetingAgendaService.loadAgenda(agendaId)).thenReturn(of(agenda));
+
+        assertThatCode(() -> service.loadSessionAgenda(agendaId))
                 .doesNotThrowAnyException();
     }
